Add files via upload
Original data file resides in google drive.
Data was constructed by the group for this app since we can't find a readily available data which is suitable for our project.
The images reside in Cloudinary which can be called via the link present in the for each item. 
Attached file is a copy for reference.